feat: [provider-ssession-bugs-003] - [Render provider-session details successfully in the website when the backing Codex session exists under the real local storage shape]
