+++ ui/scripts/verify-import-export-storybook-responsive.mjs
@@ -12,6 +12,7 @@ import {
   verifyLocalizedWorkflowActivity,
   verifyLocalizedWorkOutcomeChart,
 } from "./verify-localized-widget-storybook-responsive.mjs";
+import { verifyProviderSessionDetailSuccess as verifyProviderSessionDetailSuccessImpl } from "./verify-provider-session-storybook-responsive.mjs";
 const STORYBOOK_HOST = process.env.AGENT_FACTORY_STORYBOOK_HOST ?? "127.0.0.1";
 const STORYBOOK_PORT = process.env.AGENT_FACTORY_STORYBOOK_PORT ?? "6008";
 const STORYBOOK_URL = `http://${STORYBOOK_HOST}:${STORYBOOK_PORT}`;
@@ -418,34 +419,12 @@ export async function verifyProviderSessionDetailSuccess(
   _dialog,
   viewport,
 ) {
-  const providerSessionDetails = page.getByRole("region", {
-    name: "Selected session details",
-  });
-  await expectVisible(providerSessionDetails, "Provider-session details region");
-  await expectVisible(
-    providerSessionDetails.getByRole("heading", {
-      name: "Selected session details",
-    }),
-    "Selected session details heading",
-  );
-  await expectVisible(
-    providerSessionDetails.getByRole("heading", { name: "Source file" }),
-    "Provider-session source file heading",
-  );
-  await expectVisible(
-    providerSessionDetails.getByText(
-      "2026/05/20/rollout-2026-05-20T17-35-24-019e44f4-580e-7f32-981e-1e54ec6907d6.jsonl",
-    ),
-    "Timestamp-prefixed provider-session source path",
-  );
-  await expectVisible(
-    providerSessionDetails.getByRole("heading", { name: "Token usage" }),
-    "Provider-session token usage heading",
-  );
-  await expectNoHorizontalOverflow(
+  return verifyProviderSessionDetailSuccessImpl({
+    expectNoHorizontalOverflow,
+    expectVisible,
     page,
-    `Provider-session detail success at ${viewport.label}`,
-  );
+    viewport,
+  });
 }
 
 export async function verifyStory(browser, storyCheck, viewport) {
